+++ src/liblvr2/registration/EigenSVDPointAlign.cpp
@@ -42,32 +42,32 @@
 using namespace Eigen;
 using std::numeric_limits;
 
-using Matrix6d = Matrix<double, 6, 6>;
-using Vector6d = Matrix<double, 6, 1>;
+using Matrix6f = Matrix<float, 6, 6>;
+using Vector6f = Matrix<float, 6, 1>;
 
 namespace lvr2
 {
 
 double EigenSVDPointAlign::alignPoints(
     const PointPairVector& pairs,
-    const Vector3d& centroid_m,
-    const Vector3d& centroid_d,
-    Matrix4d& align)
+    const Vector3f& centroid_m,
+    const Vector3f& centroid_d,
+    Matrix4f& align)
 {
 /*
-    Vector3d pos, theta;
+    Vector3f pos, theta;
     matrixToPose(align, pos, theta);
     
     /// create transform matrix of the first scan
-    Matrix4d trans_m = poseToMatrix(pos, theta);
+    Matrix4f trans_m = poseToMatrix(pos, theta);
 
-    Vector3d sin, cos;
+    Vector3f sin, cos;
     sincos(theta.x(), &sin.x(), &cos.x());
     sincos(theta.y(), &sin.y(), &cos.y());
     sincos(theta.z(), &sin.z(), &cos.z());
 
     /// create matrix H
-    Matrix6d H = Matrix6d::Identity();
+    Matrix6f H = Matrix6f::Identity();
     H(0, 4) = -pos.z() * cos.x() + pos.y() * sin.x();
     H(0, 5) = pos.y() * cos.x() * cos.y() + pos.z() * cos.y() * sin.x();
     H(1, 3) = pos.z();
@@ -82,19 +82,19 @@ double EigenSVDPointAlign::alignPoints(
     H(5, 4) = cos.x();
     H(5, 5) = -cos.y() * sin.x();
 
-    #pragma omp declare reduction (+: Vector3d: omp_out=omp_out+omp_in) initializer(omp_priv=Vector3d::Zero())
-    #pragma omp declare reduction (+: Vector6d: omp_out=omp_out+omp_in) initializer(omp_priv=Vector6d::Zero())
+    #pragma omp declare reduction (+: Vector3f: omp_out=omp_out+omp_in) initializer(omp_priv=Vector3f::Zero())
+    #pragma omp declare reduction (+: Vector6f: omp_out=omp_out+omp_in) initializer(omp_priv=Vector6f::Zero())
 
-    Vector3d mid_sum = Vector3d::Zero();
-    Vector6d mz = Vector6d::Zero();
+    Vector3f mid_sum = Vector3f::Zero();
+    Vector6f mz = Vector6f::Zero();
     double xpy = 0.0, xpz = 0.0, ypz = 0.0;
     double xy = 0.0, xz = 0.0, yz = 0.0;
     double error = 0.0;
     #pragma omp parallel for reduction(+:mid_sum, xpy, xpz, ypz, xy, xz, yz, mz, error)
     for (size_t i = 0; i < pairs.size(); i++)
     {
-        Vector3d mid = (*pairs[i].second + *pairs[i].first) / 2.0;
-        Vector3d delta = *pairs[i].second - *pairs[i].first;
+        Vector3f mid = (*pairs[i].second + *pairs[i].first) / 2.0;
+        Vector3f delta = *pairs[i].second - *pairs[i].first;
 
         mid_sum += mid;
 
@@ -109,7 +109,7 @@ double EigenSVDPointAlign::alignPoints(
         yz += mid.y() * mid.z();
 
         mz.block<3, 1>(0, 0) += delta;
-        mz.block<3, 1>(3, 0) += Vector3d(-mid.z() * delta.y() + mid.y() * delta.z(),
+        mz.block<3, 1>(3, 0) += Vector3f(-mid.z() * delta.y() + mid.y() * delta.z(),
                               -mid.y() * delta.x() + mid.x() * delta.y(),
                               mid.z() * delta.x() - mid.x() * delta.z());
         
@@ -118,7 +118,7 @@ double EigenSVDPointAlign::alignPoints(
 
     error = sqrt(error / pairs.size());
 
-    Matrix6d mm = Matrix6d::Identity();
+    Matrix6f mm = Matrix6f::Identity();
     mm(0, 0) = mm(1, 1) = mm(2, 2) = pairs.size();
     mm(3, 3) = ypz;
     mm(4, 4) = xpy;
@@ -133,18 +133,18 @@ double EigenSVDPointAlign::alignPoints(
     mm(3, 5) = mm(5, 3) = -xy;
     mm(4, 5) = mm(5, 4) = -yz;
 
-    Vector6d eHat = mm.inverse() * mz;
+    Vector6f eHat = mm.inverse() * mz;
 
     /// the vector pose_d is the pose estimation of the
     // second scan = the final pose of the first scan
-    Vector6d pose_d;
+    Vector6f pose_d;
     pose_d.block<3, 1>(0, 0) = pos;
     pose_d.block<3, 1>(3, 0) = theta;
 
     pose_d -= H.inverse() * eHat;
 
     /// transform of the second scan as computed so far
-    Matrix4d trans_d = poseToMatrix(pose_d.block<3, 1>(0, 0), pose_d.block<3, 1>(3, 0));
+    Matrix4f trans_d = poseToMatrix(pose_d.block<3, 1>(0, 0), pose_d.block<3, 1>(3, 0));
 
     /// the incremental transform calculated from the absolute poses
     //  of the two scans
@@ -156,15 +156,15 @@ double EigenSVDPointAlign::alignPoints(
     double error = 0.0;
 
     // Fill H matrix
-    Matrix3d H = Matrix3d::Zero();
+    Matrix3f H = Matrix3f::Zero();
 
     // openMP seems to make this part slower
-    // #pragma omp declare reduction (+: Matrix3d: omp_out=omp_out+omp_in) initializer(omp_priv=Matrix3d::Zero())
+    // #pragma omp declare reduction (+: Matrix3f: omp_out=omp_out+omp_in) initializer(omp_priv=Matrix3f::Zero())
     // #pragma omp parallel for reduction(+:H)
     for (size_t i = 0; i < pairs.size(); i++)
     {
-        Vector3d m = *pairs[i].first - centroid_m;
-        Vector3d d = *pairs[i].second - centroid_d;
+        Vector3f m = *pairs[i].first - centroid_m;
+        Vector3f d = *pairs[i].second - centroid_d;
 
         error += (m - d).squaredNorm();
 
@@ -180,16 +180,16 @@ double EigenSVDPointAlign::alignPoints(
 
     error = sqrt(error / (double)pairs.size());
 
-    JacobiSVD<Matrix3d> svd(H, ComputeFullU | ComputeFullV);
+    JacobiSVD<Matrix3f> svd(H, ComputeFullU | ComputeFullV);
 
-    Matrix3d U = svd.matrixU();
-    Matrix3d V = svd.matrixV();
+    Matrix3f U = svd.matrixU();
+    Matrix3f V = svd.matrixV();
 
-    Matrix3d R = V * U.transpose();
+    Matrix3f R = V * U.transpose();
     align.block<3, 3>(0, 0) = R;
 
     // Calculate translation
-    Vector3d translation = centroid_d - R * centroid_m;
+    Eigen::Vector3f translation = centroid_d - R * centroid_m;
     align.block<3, 1>(0, 3) = translation;
 
     return error;
